♻️ (Week Export) - Split component into smaller reusable components to improve maintainability and reusability
