+++ src/components/WeekView.tsx
@@ -1,17 +1,23 @@
 import { Trash2Icon } from "lucide-react";
 import React, { useEffect } from "react";
 import useWeekStore from "../store/weekStore";
-import ExportDrawer from "./ExportDrawer";
 import { Button } from "./ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "./ui/card";
+import WeekExport from "./WeekExport/index";
 import WeekNavigation from "./WeekNavigation";
 import WeekSummary from "./WeekSummary";
 import WeekTable from "./WeekTable";
 
 const WeekView: React.FC = () => {
   // Récupérer les données et actions du store
-  const { currentWeekId, weeks, initializeWeek, updateDay, resetDay, resetWeek } =
-    useWeekStore();
+  const {
+    currentWeekId,
+    weeks,
+    initializeWeek,
+    updateDay,
+    resetDay,
+    resetWeek,
+  } = useWeekStore();
 
   // Initialiser la semaine au chargement du composant
   useEffect(() => {
@@ -64,7 +70,7 @@ const WeekView: React.FC = () => {
 
       {/* Options d'export */}
       <div className="flex justify-center mb-6">
-        <ExportDrawer />
+        <WeekExport />
       </div>
 
       {/* Bouton de réinitialisation - visible uniquement si des heures sont enregistrées */}
